+++ CHANGELOG.md
@@ -1,7 +1,7 @@
 2.1.0
 ====
 
-* (feature) Introduced Bundleconfiguration to set the Duration when old cron job logs should be deleted.
+*   (feature) Automatically delete old cron job logs.
 
 
 2.0.2